Let git aliases use default branch over master
Some repos I deal with don't use `master` as the default branch name.
But origin/HEAD will safely point to the default branch for the remote
(assuming the `origin` remote name is in use, but that's not a problem I
feel like dealing with today).

Note if you have an existing repo which switched away from using `master`
you may need to tell your local clone of the change:

    git remote set-head origin new-default-branch-name