Use either WiFi or Ethernet for online state
The default behaviour of systemd-networkd-wait-online is that it will
require all interfaces matched in the [Match] blocks in `/etc/systemd/network/*.network` to be online before the service reaches a successful exit. On my setup, this means both WiFi and Ethernet would need to be online. This is almost never the reality of how I use my system. Instead, I just need either of those interfaces to be online.