Freeze pump speed to 70%
The liquid heats up much more slowly than the CPU, and sustains a
temperature for much longer. Under the current pump speed configuration
this tends to mean that the longer the computer is on, the more the
pump will speed up, until it eventually reaches 100% speed (while the
liquid seems to never exceed 30°C). A single pump speed is quieter with
no significant difference in CPU temperature.