Add split audio sync script
I sometimes want to create a separate audio sync so I can stream game
audio exclusively. This allows me to do that.

When choosing audio output, select your standard audio output device if
you want to hear the sound, but don't want to stream it. Select the Null
Output if you want to stream it, but don't want to hear it. Select the
Simultaneous output (combine output) if you want to hear it AND stream
it.

https://obsproject.com/forum/resources/include-exclude-audio-sources-using-pulseaudio-linux.95/