Fix topgrade dotter command not working from ~/.dotfiles
I originally attempted this in 24e78c33f7ba28ec2d5ced7025fe362f88824336.

It seems I didn't test it. It still failed, this time with a Fish error
(which makes a lot of sense - parentheses are used for command
substitution in Fish, not for subshells). Hopefully running bash with a
specific command string will solve the issue.