Use LspAttach event for setting up nvim-navic
Following the improvements to LSP config in Neovim 0.11, some existing
configuration needs to be refactored to leverage the improvements fully.
Although the `on_attach` function was working at a basic level, it would
become difficult to work with when using the Neovim's new system for
managing language specific LSP configuration overrides.

https://neovim.io/doc/user/lsp.html#lsp-attach
https://github.com/mason-org/mason-lspconfig.nvim/issues/545#issuecomment-2906711628